dev-requirements.txt
ipython-7.4.0
django-extensions-2.1.6
django-debug-toolbar-1.11